+++ src/components/TaskBoard/TaskBoard.tsx
@@ -1,36 +1,25 @@
-import { defineComponent, ref } from 'vue';
+// TaskBoard.tsx
+import { defineComponent } from 'vue';
 import TaskColumn from './TaskColumn';
-import { initialTasks, getTasksByState } from '../../data/tasks';
 import { TaskState } from '../../model/task';
+import { useTasks } from '../tasks/useTasks'; // جديد
 import '../../Styles/TaskBoard.css';
 
 export default defineComponent({
   name: 'TaskBoard',
   components: { TaskColumn },
   setup() {
-    const tasks = ref(initialTasks);
+    const {
+      getTasksByState,
+      handleTaskDrop,
+      handleDragStart
+    } = useTasks();
 
-    const filterTasksByState = (state: TaskState) => {
-      return getTasksByState(tasks.value, state);
-    };
-
-
-    const handleTaskDrop = ({ taskId, newState }: { taskId: number, newState: TaskState }) => {
-      const task = tasks.value.find(t => t.id === taskId);
-      if (task) task.state = newState;
-    };
-    const draggedTaskId = ref<number | null>(null);
-
-    const handleDragStart = (taskId: number) => {
-      draggedTaskId.value = taskId;
-      document.querySelectorAll('.task-column').forEach(el => {
-        el.setAttribute('data-dragged', taskId.toString());
-      });
-    };
     return {
       TaskState,
-      getTasksByState: filterTasksByState,
-      handleTaskDrop, handleDragStart
+      getTasksByState,
+      handleTaskDrop,
+      handleDragStart
     };
   },
   render() {
